+++ app/src/main/java/com/vultisig/wallet/app/activity/MainActivity.kt
@@ -16,22 +16,24 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.core.view.WindowCompat
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.compose.rememberNavController
 import com.google.android.play.core.appupdate.AppUpdateManager
 import com.google.android.play.core.appupdate.AppUpdateOptions
 import com.google.android.play.core.install.model.AppUpdateType
-import com.google.android.play.core.install.model.UpdateAvailability
 import com.vultisig.wallet.ui.components.BiometryAuthScreen
 import com.vultisig.wallet.ui.navigation.SetupNavGraph
 import com.vultisig.wallet.ui.navigation.route
 import com.vultisig.wallet.ui.theme.OnBoardingComposeTheme
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 
 @AndroidEntryPoint
 internal class MainActivity : AppCompatActivity() {
 
     private val mainViewModel: MainViewModel by viewModels<MainViewModel>()
+
+    @Inject
+    internal lateinit var appUpdateManager: AppUpdateManager
 
     override fun onCreate(savedInstanceState: Bundle?) {
         installSplashScreen()
@@ -43,7 +45,15 @@ internal class MainActivity : AppCompatActivity() {
 
         WindowCompat.setDecorFitsSystemWindows(window, false)
 
-        mainViewModel.checkUpdates(savedInstanceState)
+        mainViewModel.checkUpdates { appUpdateInfo ->
+            appUpdateManager.startUpdateFlowForResult(
+                appUpdateInfo,
+                this@MainActivity,
+                AppUpdateOptions.newBuilder(AppUpdateType.IMMEDIATE)
+                    .setAllowAssetPackDeletion(true)
+                    .build(), 0
+            )
+        }
 
         setContent {
             OnBoardingComposeTheme {
